+++ src/actions/index.js
@@ -68,35 +68,10 @@ export {
 
 export { getUserSchema } from './userschema/userschema';
 
-export { GET_SUBSCRIBERS, getSubscribers } from './newsletter/get_subscribers';
 export {
-  POST_SUBSCRIBERS,
-  subscribeNewsletter,
-  confirmSubscribe,
-} from './newsletter/post_subscribers';
-export {
-  POST_UNSUBSCRIBERS,
-  unsubscribeNewsletter,
-  confirmUnsubscribe,
-} from './newsletter/post_subscribers';
-export {
-  POST_SUBSCRIBE_TO_EVENT,
-  POST_UNSUBSCRIBE_TO_EVENT,
-  POST_CONFIRM_SUBSCRIBE_TO_EVENT,
-  POST_CONFIRM_UNSUBSCRIBE_TO_EVENT,
-  subscribeToEvent,
-  confirmSubscribeToEvent,
-  unsubscribeToEvent,
-  confirmUnsubscribeToEvent,
-} from './subscribe_to_event/post_subscribe_to_event';
-export {
-  POST_SUBSCRIBE_TO_NEWS,
-  POST_UNSUBSCRIBE_TO_NEWS,
-  subscribeToNews,
-  confirmSubscribeToNews,
-  unsubscribeToNews,
-  confirmUnsubscribeToNews,
-} from './subscribe_to_news/post_subscribe_to_news';
+  GET_NEWSLETTER,
+  getNewsletterSubscriber,
+} from './newsletter/get_newsletter_subscribers';
 
 export {
   POST_SUBSCRIBE_TO,
