+++ src/actions/newsletter/get_newsletter_subscribers.js
@@ -2,16 +2,16 @@
  * Get subscribers.
  * @module actions/newsletter
  */
-export const GET_SUBSCRIBERS = 'GET_SUBSCRIBERS';
+export const GET_NEWSLETTER = 'GET_NEWSLETTER';
 
 /**
  * Get subscribers.
- * @function GetSubscribers
+ * @function GetNewsletterSubscriber
  * @returns {Object} Get extra items action.
  */
-export function getSubscribers() {
+export function getNewsletterSubscriber() {
   return {
-    type: GET_SUBSCRIBERS,
+    type: GET_NEWSLETTER,
     request: {
       op: 'get',
       path: `/@newsletter-subscribers`,
