+++ src/components/CLMSProfileView/index.js
@@ -1,7 +1,4 @@
 export CLMSProfileView from './CLMSProfileView';
 export CLMSUserProfileView from './CLMSUserProfileView';
 export CLMSApiTokensView from './CLMSApiTokensView';
-export CLMSNewsletterView from './CLMSNewsletterView';
-export CLMSSubscribeToEventsView from './CLMSSubscribeToEventsView';
-export CLMSSubscribeToNewsView from './CLMSSubscribeToNewsView';
 export CLMSNewsletterSubscriberView from './CLMSNewsletterSubscriberView';
