+++ src/components/CLMSSubscriptionView/index.js
@@ -1,3 +1,3 @@
 export SubscriptionView from './SubscriptionView';
 export ConfirmSubscriptionView from './ConfirmSubscriptionView';
-export {AVAILABLE_SUBSCRIPTIONS} from './subscription_utils';
+export { AVAILABLE_SUBSCRIPTIONS } from './subscription_utils';
