+++ src/reducers/index.js
@@ -13,10 +13,8 @@ import { registryReducer } from './registry/registry';
 // import defaultReducers from '@plone/volto/reducers';
 import tokens from './tokens/tokens';
 import { userschemaReducer } from './userschema/userschema';
-import { newsletterReducer } from './newsletter/newsletter_reducer';
-import { subscribeToEventReducer } from './subscribe_to_event/subscribe_to_event_reducer';
-import { subscribeToNewsReducer } from './subscribe_to_news/subscribe_to_news_reducer';
-import { newsletterSubscribersReducer } from './newsletter/subscriber_reducer';
+// import { newsletterReducer } from './newsletter/newsletter_reducer';
+import { newsletterSubscribersReducer } from './newsletter/get_newsletter_reducer';
 import { subscribeToReducer } from './subscription/subscription_reducer';
 
 /**
@@ -38,9 +36,7 @@ const reducers = {
   geonetwork_importation: importGeonetworkReducer,
   userschema: userschemaReducer,
   datasetsByUid: datasetsByUidReducer,
-  subscribe_to_newsletter: newsletterReducer,
-  subscribe_to_event: subscribeToEventReducer,
-  subscribe_to_news: subscribeToNewsReducer,
+  // subscribe_to_newsletter: newsletterReducer,
   newsletter_subscribers: newsletterSubscribersReducer,
   subscription: subscribeToReducer,
 };
